# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open Sky Holidays API",
        "description": "API documentation for Open Sky Holidays",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/why-choose-sections/{whyChooseSection}": {
            "get": {
                "tags": [
                    "Why Choose Us"
                ],
                "summary": "Get single Why Choose Us section details",
                "description": "Retrieves single Why Choose Us card details by ID.",
                "operationId": "7877309fabf7030512bb440f76324f93",
                "parameters": [
                    {
                        "name": "whyChooseSection",
                        "in": "path",
                        "description": "Why Choose Section ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Why choose section details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/WhyChooseSection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "WhyChooseSection": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Worldwide Coverage"
                    },
                    "description": {
                        "type": "string",
                        "example": "Explore domestic and international destinations with complete planning and trusted travel support."
                    },
                    "icon": {
                        "type": "string",
                        "example": "fa-solid fa-earth-americas",
                        "nullable": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}